requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.24.0
pydantic==2.10.6
openai==1.91.0
python-dotenv==1.0.0
httpx==0.28.1
aiofiles==23.2.1
asyncio==3.4.3
streamlit==1.29.0
websockets==12.0
pandas==2.1.4